Move SpriteFrames to its own file in the resources folder
