+++ scene/resources/sprite_frames.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  animated_sprite_2d.h                                                 */
+/*  sprite_frames.h                                                      */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,89 +28,75 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef ANIMATED_SPRITE_2D_H
-#define ANIMATED_SPRITE_2D_H
+#ifndef SPRITE_FRAMES_H
+#define SPRITE_FRAMES_H
 
-#include "scene/2d/node_2d.h"
-#include "scene/resources/sprite_frames.h"
 #include "scene/resources/texture.h"
 
-class AnimatedSprite2D : public Node2D {
-	GDCLASS(AnimatedSprite2D, Node2D);
+class SpriteFrames : public Resource {
+	GDCLASS(SpriteFrames, Resource);
 
-	Ref<SpriteFrames> frames;
-	bool playing = false;
-	bool backwards = false;
-	StringName animation = "default";
-	int frame = 0;
-	float speed_scale = 1.0f;
+	struct Anim {
+		float speed = 5.0;
+		bool loop = true;
+		Vector<Ref<Texture2D>> frames;
+	};
 
-	bool centered = true;
-	Point2 offset;
+	Map<StringName, Anim> animations;
 
-	bool is_over = false;
-	float timeout = 0.0;
+	Array _get_frames() const;
+	void _set_frames(const Array &p_frames);
 
-	bool hflip = false;
-	bool vflip = false;
+	Array _get_animations() const;
+	void _set_animations(const Array &p_animations);
 
-	void _res_changed();
-
-	float _get_frame_duration();
-	void _reset_timeout();
-	void _set_playing(bool p_playing);
-	bool _is_playing() const;
-	Rect2 _get_rect() const;
+	Vector<String> _get_animation_list() const;
 
 protected:
 	static void _bind_methods();
-	void _notification(int p_what);
-	virtual void _validate_property(PropertyInfo &property) const override;
 
 public:
-#ifdef TOOLS_ENABLED
-	virtual Dictionary _edit_get_state() const override;
-	virtual void _edit_set_state(const Dictionary &p_state) override;
+	void add_animation(const StringName &p_anim);
+	bool has_animation(const StringName &p_anim) const;
+	void remove_animation(const StringName &p_anim);
+	void rename_animation(const StringName &p_prev, const StringName &p_next);
 
-	virtual void _edit_set_pivot(const Point2 &p_pivot) override;
-	virtual Point2 _edit_get_pivot() const override;
-	virtual bool _edit_use_pivot() const override;
-	virtual Rect2 _edit_get_rect() const override;
-	virtual bool _edit_use_rect() const override;
-#endif
+	void get_animation_list(List<StringName> *r_animations) const;
+	Vector<String> get_animation_names() const;
 
-	virtual Rect2 get_anchorable_rect() const override;
+	void set_animation_speed(const StringName &p_anim, float p_fps);
+	float get_animation_speed(const StringName &p_anim) const;
 
-	void set_sprite_frames(const Ref<SpriteFrames> &p_frames);
-	Ref<SpriteFrames> get_sprite_frames() const;
+	void set_animation_loop(const StringName &p_anim, bool p_loop);
+	bool get_animation_loop(const StringName &p_anim) const;
 
-	void play(const StringName &p_animation = StringName(), const bool p_backwards = false);
-	void stop();
-	bool is_playing() const;
+	void add_frame(const StringName &p_anim, const Ref<Texture2D> &p_frame, int p_at_pos = -1);
+	int get_frame_count(const StringName &p_anim) const;
+	_FORCE_INLINE_ Ref<Texture2D> get_frame(const StringName &p_anim, int p_idx) const {
+		const Map<StringName, Anim>::Element *E = animations.find(p_anim);
+		ERR_FAIL_COND_V_MSG(!E, Ref<Texture2D>(), "Animation '" + String(p_anim) + "' doesn't exist.");
+		ERR_FAIL_COND_V(p_idx < 0, Ref<Texture2D>());
+		if (p_idx >= E->get().frames.size()) {
+			return Ref<Texture2D>();
+		}
 
-	void set_animation(const StringName &p_animation);
-	StringName get_animation() const;
+		return E->get().frames[p_idx];
+	}
 
-	void set_frame(int p_frame);
-	int get_frame() const;
+	void set_frame(const StringName &p_anim, int p_idx, const Ref<Texture2D> &p_frame) {
+		Map<StringName, Anim>::Element *E = animations.find(p_anim);
+		ERR_FAIL_COND_MSG(!E, "Animation '" + String(p_anim) + "' doesn't exist.");
+		ERR_FAIL_COND(p_idx < 0);
+		if (p_idx >= E->get().frames.size()) {
+			return;
+		}
+		E->get().frames.write[p_idx] = p_frame;
+	}
+	void remove_frame(const StringName &p_anim, int p_idx);
+	void clear(const StringName &p_anim);
+	void clear_all();
 
-	void set_speed_scale(float p_speed_scale);
-	float get_speed_scale() const;
-
-	void set_centered(bool p_center);
-	bool is_centered() const;
-
-	void set_offset(const Point2 &p_offset);
-	Point2 get_offset() const;
-
-	void set_flip_h(bool p_flip);
-	bool is_flipped_h() const;
-
-	void set_flip_v(bool p_flip);
-	bool is_flipped_v() const;
-
-	virtual String get_configuration_warning() const override;
-	AnimatedSprite2D();
+	SpriteFrames();
 };
 
-#endif // ANIMATED_SPRITE_H
+#endif // SPRITE_FRAMES_H
